:recycle: [#117] Refactored Validsign auth
