+++ src/bptl/work_units/valid_sign/authentication.py
@@ -1,44 +1,7 @@
-from django.utils.crypto import constant_time_compare
-from django.utils.translation import gettext_lazy as _
-
-from rest_framework import exceptions
-from rest_framework.authentication import BaseAuthentication, get_authorization_header
-from rest_framework.request import Request
-
-from bptl.accounts.models import User
-
+from ..authentication import WebhookAuthentication
 from .models import ValidSignConfiguration
 
 
-class ValidSignAuthentication(BaseAuthentication):
-
-    # Bulk taken from rest_framework.authentication.BasicAuthentication
-
-    @staticmethod
-    def _get_auth_key() -> str:
-        config = ValidSignConfiguration.get_solo()
-        return config.auth_key
-
-    def authenticate(self, request: Request):
-        auth = get_authorization_header(request).split()
-        if not auth or auth[0].lower() != b"basic":
-            return None
-
-        if len(auth) == 1:
-            msg = _("Invalid basic header. No credentials provided.")
-            raise exceptions.AuthenticationFailed(msg)
-        elif len(auth) > 2:
-            msg = _(
-                "Invalid basic header. Credentials string should not contain spaces."
-            )
-            raise exceptions.AuthenticationFailed(msg)
-
-        # extract the key from the header
-        key = auth[1].decode()
-
-        if not constant_time_compare(key, self._get_auth_key()):
-            return None
-
-        # User does not actually exist in the database!
-        user = User(username="ValidSign API")
-        return (user, None)
+class ValidSignAuthentication(WebhookAuthentication):
+    config_class = ValidSignConfiguration
+    application_name = "ValidSign API"
